feat(packages/sui-mono): Rewrite Prmpter Manager and change name to use kebab case
